tidy up _app move logic to hooks and add component files for vega connect button and app loader
